requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
alabaster=0.7.10=py27h9dd7d6e_0
anaconda=5.0.0=py27h58f0b96_0
anaconda-client=1.6.5=py27hc13fba8_0
anaconda-navigator=1.6.8=py27h308d521_1
anaconda-project=0.8.0=py27h9e3d455_0
appnope=0.1.0=py27hb466136_0
appscript=1.0.1=py27h451298e_1
asn1crypto=0.22.0=py27h61af4a7_1
astroid=1.5.3=py27h96f3fd4_0
astropy=2.0.2=py27h87cc2bd_4
babel=2.5.0=py27h7311c9e_0
backports=1.0=py27hb4f9756_1
backports.functools_lru_cache=1.4=py27h2aca819_1
backports.shutil_get_terminal_size=1.0.0=py27hc9115de_2
backports_abc=0.5=py27h6972548_0
beautifulsoup4=4.6.0=py27h9416283_1
bitarray=0.8.1=py27hd5bfd95_0
bkcharts=0.2=py27haafc882_0
blaze=0.11.3=py27hb49378a_0
bleach=2.0.0=py27ha7d1710_0
bokeh=0.12.7=py27h22b55c5_1
boto=2.48.0=py27hacdd0fd_1
bottleneck=1.2.1=py27h71f98a3_0
ca-certificates=2017.08.26=ha1e5d58_0
cdecimal=2.3=py27hf5d9fd9_1
certifi=2017.7.27.1=py27h482ffc0_0
cffi=1.10.0=py27haac214c_1
chardet=3.0.4=py27h2842e91_1
click=6.7=py27h2b86a94_0
cloudpickle=0.4.0=py27h665dddb_0
clyent=1.2.2=py27hc0ae608_0
colorama=0.3.9=py27hbbe92b6_0
configparser=3.5.0=py27hc7edf1b_0
contextlib2=0.5.5=py27h9cb85f4_0
cryptography=2.0.3=py27hab69567_1
curl=7.55.1=h7601780_3
cycler=0.10.0=py27hfc73c78_0
cython=0.26.1=py27h6a053f9_0
cytoolz=0.8.2=py27ha56cda1_0
dask=0.15.2=py27h32f4218_0
dask-core=0.15.2=py27h5003087_0
datashape=0.5.4=py27hd6a1745_0
dbus=1.10.22=h50d9ad6_0
decorator=4.1.2=py27h9f877ea_0
distributed=1.18.3=py27h81d79ea_0
docutils=0.14=py27h0befae3_0
entrypoints=0.2.3=py27hd680fb1_2
enum34=1.1.6=py27hf475452_1
et_xmlfile=1.0.1=py27hc42f929_0
expat=2.2.4=h8f26bf8_1
fastcache=1.0.2=py27hc4635c7_0
filelock=2.0.12=py27h72fe922_0
flask=0.12.2=py27h3ac5568_0
flask-cors=3.0.3=py27h13db576_0
freetype=2.8=h143eb01_0
funcsigs=1.0.2=py27hb9f6266_0
functools32=3.2.3.2=py27h8ceab06_1
futures=3.1.1=py27hb02a37a_0
get_terminal_size=1.0.0=h7520d66_0
gettext=0.19.8.1=hb0f4f8b_2
gevent=1.2.2=py27hc02608c_0
glib=2.53.6=ha08cb78_1
glob2=0.5=py27h2f8fe13_0
gmp=6.1.2=h4a9834d_0
gmpy2=2.0.8=py27h7e2fca4_1
greenlet=0.4.12=py27h081ed54_0
grin=1.2.1=py27hc43e5f3_1
h5py=2.7.0=py27h217cc45_1
hdf5=1.10.1=h6090a45_0
heapdict=1.0.0=py27hb5e74ad_0
html5lib=0.999999999=py27hec7e2bc_0
icu=58.2=hea21ae5_0
idna=2.6=py27hedea723_1
imageio=2.2.0=py27h37746d9_0
imagesize=0.7.1=py27h4f7bcc8_0
intel-openmp=2018.0.0=hdd0ccc9_7
ipaddress=1.0.18=py27h5b9a5b9_0
ipykernel=4.6.1=py27h1e70a78_0
ipython=5.4.1=py27h2b3d779_1
ipython_genutils=0.2.0=py27h8b9a179_0
ipywidgets=7.0.0=py27h3e52029_0
isort=4.2.15=py27h5bf637f_0
itsdangerous=0.24=py27h3948ded_1
jbig=2.1=h4d881f8_0
jdcal=1.3=py27hfeaf94f_0
jedi=0.10.2=py27hbb5dc62_0
jinja2=2.9.6=py27h92590e2_1
jpeg=9b=haccd157_1
jsonschema=2.6.0=py27hd9b497e_0
jupyter=1.0.0=py27hec63c99_0
jupyter_client=5.1.0=py27hfaf569a_0
jupyter_console=5.2.0=py27h9702a86_1
jupyter_core=4.3.0=py27hd5161ba_0
jupyterlab=0.27.0=py27h25d4955_2
jupyterlab_launcher=0.4.0=py27he518b91_0
lazy-object-proxy=1.3.1=py27h712ce3f_0
libcxx=4.0.1=h579ed51_0
libcxxabi=4.0.1=hebd6815_0
libedit=3.1=hb4e282d_0
libffi=3.2.1=hd939716_3
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=h99df5da_5
libpng=1.6.32=hce72d48_2
libsodium=1.0.13=hba5e272_2
libssh2=1.8.0=h1218725_2
libtiff=4.0.8=h8cd0352_9
libxml2=2.9.4=hbd0960b_5
libxslt=1.1.29=h95a2935_5
llvmlite=0.20.0=py27_0
locket=0.2.0=py27ha10513d_1
lxml=3.8.0=py27h037e1f5_0
markupsafe=1.0=py27hd3c86fa_1
matplotlib=2.0.2=py27h2e09848_1
mccabe=0.6.1=py27h1f69e8d_0
mistune=0.7.4=py27h1658d75_0
mkl=2018.0.0=hc285769_4
mkl-service=1.1.2=py27hcba3029_4
mpc=1.0.3=hc455b36_4
mpfr=3.1.5=h7fa3772_1
mpmath=0.19=py27h09cdc99_2
msgpack-python=0.4.8=py27h635ded4_0
multipledispatch=0.4.9=py27h10993aa_0
navigator-updater=0.1.0=py27ha63e0b4_0
nbconvert=5.3.1=py27h6455e4c_0
nbformat=4.4.0=py27hddc86d0_0
ncurses=6.0=ha932d30_1
networkx=1.11=py27h0a72f81_0
nltk=3.2.4=py27h1626047_0
nose=1.3.7=py27h2ee3cb8_2
notebook=5.0.0=py27h5f5981d_2
numba=0.35.0=np113py27_6
numexpr=2.6.2=py27h99e4b4f_1
numpy=1.13.1=py27hd567e90_2
numpydoc=0.7.0=py27h022f19e_0
odo=0.5.1=py27h992a9f7_0
olefile=0.44=py27h73ba740_0
openpyxl=2.4.8=py27h70e7ed9_1
openssl=1.0.2l=h57f3a61_2
packaging=16.8=py27h24b219a_0
pandas=0.20.3=py27h2a3d8ca_2
pandoc=1.19.2.1=ha5e8f32_1
pandocfilters=1.4.2=py27hed78c4e_1
partd=0.3.8=py27h7560dbf_0
path.py=10.3.1=py27h5e25276_0
pathlib2=2.3.0=py27he09da1e_0
patsy=0.4.1=py27h40ed276_0
pcre=8.41=h29eefc5_0
pep8=1.7.0=py27hff3397c_0
pexpect=4.2.1=py27hc4e4961_0
pickleshare=0.7.4=py27h37e3d41_0
pillow=4.2.1=py27h2cf1d5f_0
pip=9.0.1=py27h5494384_2
pkginfo=1.4.1=py27ha9221e7_0
ply=3.10=py27h6279b8a_0
prompt_toolkit=1.0.15=py27h4a7b9c2_0
psutil=5.2.2=py27h59435ec_0
ptyprocess=0.5.2=py27h70f6364_0
py=1.4.34=py27h917aa57_1
pyaudio=0.2.7=py27h3777516_1
pycodestyle=2.3.1=py27h5b634e0_0
pycosat=0.6.2=py27h085d4cc_0
pycparser=2.18=py27h0d28d88_1
pycrypto=2.6.1=py27h4efa152_1
pycurl=7.43.0=py27h398a7fe_3
pyflakes=1.5.0=py27h7dfcaa9_1
pygments=2.2.0=py27h1a556bb_0
pylint=1.7.2=py27h718c7e7_0
pyodbc=4.0.17=py27hc9de18c_0
pyopenssl=17.2.0=py27h732fe57_0
pyparsing=2.2.0=py27h5bb6aaf_0
pyqt=5.6.0=py27hf21fe59_6
pysocks=1.6.7=py27h1cff6a6_1
pytables=3.4.2=py27h6dea595_1
pytest=3.2.1=py27h173df03_1
python=2.7.13=h32f5f24_13
python-dateutil=2.6.1=py27hd56c96b_1
python.app=2=py27h48d88ae_5
pytz=2017.2=py27hb891d23_1
pywavelets=0.5.2=py27hd99e88a_0
pyyaml=3.12=py27ha7932d0_1
pyzmq=16.0.2=py27he61c07e_2
qt=5.6.2=h9975529_14
qtawesome=0.4.4=py27hdeb2f59_0
qtconsole=4.3.1=py27hdc90b4f_0
qtpy=1.3.1=py27h39159f8_0
readline=7.0=h81b24a6_3
requests=2.18.4=py27h9b2b37c_1
rope=0.10.5=py27he855d02_0
ruamel_yaml=0.11.14=py27h31666c4_2
scandir=1.5=py27h77d1c80_1
scikit-image=0.13.0=py27h03e84e1_1
scikit-learn=0.19.0=py27h54286d7_2
scipy=0.19.1=py27hf01dd8f_3
seaborn=0.8.0=py27h92884e4_0
setuptools=36.5.0=py27h2a45cec_0
simplegeneric=0.8.1=py27h6db5e31_0
singledispatch=3.4.0.3=py27he22c18d_0
sip=4.18.1=py27h6300f65_2
six=1.10.0=py27h47fc262_1
snowballstemmer=1.2.1=py27h68ac032_0
sortedcollections=0.5.3=py27h8094be4_0
sortedcontainers=1.5.7=py27h322dbbf_0
sphinx=1.6.3=py27h11269f0_0
sphinxcontrib=1.0=py27hd2ed746_1
sphinxcontrib-websupport=1.0.1=py27h857890b_1
spyder=3.2.3=py27h7402f24_0
sqlalchemy=1.1.13=py27hcbc9ed3_0
sqlite=3.20.1=h900c3b0_1
ssl_match_hostname=3.5.0.1=py27h8780752_2
statsmodels=0.8.0=py27h6d68dbf_0
subprocess32=3.2.7=py27h24b2887_0
sympy=1.1.1=py27hce55102_0
tblib=1.3.2=py27ha684fc4_0
terminado=0.6=py27he40bf16_0
testpath=0.3.1=py27h72d81a5_0
tk=8.6.7=hcdce994_1
toolz=0.8.2=py27h27228c4_0
tornado=4.5.2=py27h29aec9e_0
traitlets=4.3.2=py27hcf08151_0
typing=3.6.2=py27h646fea0_0
unicodecsv=0.14.1=py27h170f95c_0
unidecode=0.04.21=py27_0
unixodbc=2.3.4=h4cb4dde_1
urllib3=1.22=py27hc3787e9_0
wcwidth=0.1.7=py27h817c265_0
webencodings=0.5.1=py27h19a9f58_1
werkzeug=0.12.2=py27hcac71f8_0
wheel=0.29.0=py27h84bd1c0_1
widgetsnbextension=3.0.2=py27h56f70de_1
wrapt=1.10.11=py27hd341262_0
xlrd=1.1.0=py27hbd41ed1_1
xlsxwriter=0.9.8=py27h1633c25_0
xlwings=0.11.4=py27h4d78f01_0
xlwt=1.2.0=py27hbeec4ae_0
xz=5.2.3=ha24016e_1
yaml=0.1.7=hff548bb_1
zeromq=4.2.2=h131e0f7_1
zict=0.1.2=py27h1c86053_0
zlib=1.2.11=h60db283_1
